Move WeightInfo, Config, Event into Asset/Checkpoint pallet.
