+++ pallets/asset/Cargo.toml
@@ -8,7 +8,6 @@ edition = "2021"
 # Common
 polymesh-primitives = { workspace = true, default-features = false }
 polymesh-primitives-derive = { workspace = true, default-features = false }
-polymesh-common-utilities = { workspace = true, default-features = false }
 
 # Our Pallets
 pallet-base = { workspace = true, default-features = false  }
@@ -57,7 +56,6 @@ std = [
     "pallet-portfolio/std",
     "pallet-timestamp/std",
     "pallet-external-agents/std",
-    "polymesh-common-utilities/std",
     "polymesh-primitives/std",
     "polymesh-runtime-common/std",
     "rustc-hex/std",
@@ -73,6 +71,5 @@ std = [
 ]
 runtime-benchmarks = [
     "frame-benchmarking",
-    "polymesh-common-utilities/runtime-benchmarks",
     "pallet-identity/runtime-benchmarks",
 ]
