+++ pallets/common/src/traits/mod.rs
@@ -13,5 +13,3 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-pub mod asset;
-pub mod checkpoint;
