+++ pallets/runtime/tests/Cargo.toml
@@ -34,7 +34,6 @@ pallet-sudo = { workspace = true, default-features = false }
 pallet-transaction-payment = { workspace = true, default-features = false }
 pallet-treasury = { workspace = true, default-features = false }
 pallet-utility = { workspace = true, default-features = false }
-polymesh-common-utilities = { workspace = true, default-features = false }
 polymesh-contracts = { workspace = true, default-features = false }
 polymesh-primitives = { workspace = true, default-features = false }
 polymesh-runtime-common = { workspace = true, default-features = false, features = ["testing"] }
@@ -108,7 +107,7 @@ sp-tracing = { version = "6.0.0", default-features = false, features = ["std"] }
 [features]
 default = ["std", "testing", "equalize"]
 
-default_identity = [ "polymesh-common-utilities/default_identity" ]
+default_identity = []
 
 testing = []
 equalize = []
@@ -160,7 +159,6 @@ std = [
     "pallet-transaction-payment/std",
     "pallet-treasury/std",
     "pallet-sudo/std",
-    "polymesh-common-utilities/std",
     "polymesh-primitives/std",
     "polymesh-runtime-common/std",
     "polymesh-contracts/std",
