+++ pallets/settlement/src/lib.rs
@@ -70,7 +70,6 @@ use sp_std::vec;
 
 use pallet_asset::MandatoryMediators;
 use pallet_base::{ensure_string_limited, try_next_post};
-use polymesh_common_utilities::traits::asset;
 use polymesh_primitives::asset::AssetId;
 use polymesh_primitives::constants::queue_priority::SETTLEMENT_INSTRUCTION_EXECUTION_PRIORITY;
 use polymesh_primitives::settlement::{
@@ -400,7 +399,7 @@ pub trait WeightInfo {
 }
 
 pub trait Config:
-    asset::Config
+    pallet_asset::Config
     + frame_system::Config
     + pallet_compliance_manager::Config
     + frame_system::Config
